Separate WendyAgentError and group APIs by visibility
Keeping error and lifecycle types in their own files makes the renamed
library surface easier to scan as it grows. Grouping members by access
level also makes the intended API boundary obvious and keeps the new
Wendy-prefixed types consistent to navigate.
