+++ swift/Sources/WendyAgent/WendyAgentError.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+public enum WendyAgentError: LocalizedError, Sendable {
+    case stoppedDuringStartup
+
+    public var errorDescription: String? {
+        switch self {
+        case .stoppedDuringStartup:
+            "WendyAgent stopped before startup completed."
+        }
+    }
+}
